requirements.txt
bokeh==3.4.3
pandas==2.2.2
openpyxl==3.1.5
pystray==0.19.5
pillow==10.4.0
python-dateutil==2.9.0.post0
pyrfc>=3.3
